Homework 17.1 ready
