+++ 17.1_main.py
@@ -6,11 +6,13 @@ if __name__ == '__main__':
         product_invalid = Product("Бракованный товар", "Неверное количество", 1000.0, 0)
     except ValueError as e:
         print(
-            "Возникла ошибка ValueError прерывающая работу программы при попытке добавить продукт с нулевым количеством")
+            f"Возникла ошибка ValueError прерывающая работу программы при попытке добавить продукт с "
+            f"нулевым количеством, ошибка {e}")
     else:
         print("Не возникла ошибка ValueError при попытке добавить продукт с нулевым количеством")
 
-    product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
+    product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера",
+                       180000.0, 5)
     product2 = Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
     product3 = Product("Xiaomi Redmi Note 11", "1024GB, Синий", 31000.0, 14)
