+++ htmlcov/function_index.html
@@ -55,7 +55,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.1">coverage.py v7.9.1</a>,
-            created at 2025-07-11 15:07 +0500
+            created at 2025-07-11 15:12 +0500
         </p>
     </div>
 </header>
@@ -385,28 +385,36 @@
                 <td class="right" data-ratio="2 2">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html#t27">tests\test_category.py</a></td>
-                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html#t27"><data value='test_exceptions'>test_exceptions</data></a></td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html">tests\test_category.py</a></td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html"><data value=''><span class='no-noun'>(no function)</span></data></a></td>
+                <td>5</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="5 5">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html#t4">tests\test_exceptions.py</a></td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html#t4"><data value='test_exceptions'>test_exceptions</data></a></td>
                 <td>6</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="6 6">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html#t36">tests\test_category.py</a></td>
-                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html#t36"><data value='test_null'>test_null</data></a></td>
-                <td>7</td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html#t13">tests\test_exceptions.py</a></td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html#t13"><data value='test_null'>test_null</data></a></td>
+                <td>6</td>
                 <td>1</td>
                 <td>0</td>
-                <td class="right" data-ratio="6 7">86%</td>
+                <td class="right" data-ratio="5 6">83%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html">tests\test_category.py</a></td>
-                <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html"><data value=''><span class='no-noun'>(no function)</span></data></a></td>
-                <td>7</td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html">tests\test_exceptions.py</a></td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html"><data value=''><span class='no-noun'>(no function)</span></data></a></td>
+                <td>4</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="7 7">100%</td>
+                <td class="right" data-ratio="4 4">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_a44f0ac069e85531_test_lawngrass_py.html#t4">tests\test_lawngrass.py</a></td>
@@ -541,10 +549,10 @@
             <tr class="total">
                 <td class="name left">Total</td>
                 <td class="name left">&nbsp;</td>
-                <td>252</td>
+                <td>253</td>
                 <td>25</td>
                 <td>0</td>
-                <td class="right" data-ratio="227 252">90%</td>
+                <td class="right" data-ratio="228 253">90%</td>
             </tr>
         </tfoot>
     </table>
@@ -556,7 +564,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.1">coverage.py v7.9.1</a>,
-            created at 2025-07-11 15:07 +0500
+            created at 2025-07-11 15:12 +0500
         </p>
     </div>
     <aside class="hidden">
