+++ htmlcov/index.html
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.1">coverage.py v7.9.1</a>,
-            created at 2025-07-11 15:07 +0500
+            created at 2025-07-11 15:12 +0500
         </p>
     </div>
 </header>
@@ -142,10 +142,17 @@
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_a44f0ac069e85531_test_category_py.html">tests\test_category.py</a></td>
-                <td>31</td>
+                <td>16</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="16 16">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_a44f0ac069e85531_test_exceptions_py.html">tests\test_exceptions.py</a></td>
+                <td>16</td>
                 <td>1</td>
                 <td>0</td>
-                <td class="right" data-ratio="30 31">97%</td>
+                <td class="right" data-ratio="15 16">94%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_a44f0ac069e85531_test_lawngrass_py.html">tests\test_lawngrass.py</a></td>
@@ -179,10 +186,10 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>252</td>
+                <td>253</td>
                 <td>25</td>
                 <td>0</td>
-                <td class="right" data-ratio="227 252">90%</td>
+                <td class="right" data-ratio="228 253">90%</td>
             </tr>
         </tfoot>
     </table>
@@ -194,7 +201,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.1">coverage.py v7.9.1</a>,
-            created at 2025-07-11 15:07 +0500
+            created at 2025-07-11 15:12 +0500
         </p>
     </div>
     <aside class="hidden">
