+++ src/exceptions.py
@@ -1,4 +1,3 @@
 class ZeroRuntimeProduct(Exception):
-    def __init__(self, message = None):
+    def __init__(self, message=None):
         super().__init__(self, message)
-
