+++ tests/conftest.py
@@ -48,7 +48,7 @@ def category_wo_products():
         name="Смартфоны",
         description="Смартфоны, как средство не только коммуникации, "
         "но и получение дополнительных функций для удобства жизни",
-        )
+    )
 
 
 @pytest.fixture
